Detect system path rather than try to maintain it.
Splitting the zsh config into two files (.zshenv and .zshrc) caused
a problem where blackrose was setting the path after when .zshenv was
sourced (when the system path was captured) but before .zshrc was
sourced (when directories were added to the path using the previously
captured path).  Now the path setup detects the system path
independently each time it is called by stripping of the before/after
paths.  It also checks to make sure the same before/after path is not
added multiple times.