Change git push behavior to upstream.
The upstream push behavior means that when "git push" is given without
arguments, it pushes to the remote branch that is being tracked (what is
synced when "git pull" is called).  The (current) default behavior is to
push all tracked branches, which I always found a bit annoying.

This change supresses a warning message with the latest update from git
and prepares for an upcoming change in future git distributions.