Change backup script to split large backup files.
Prevents any single backup file from being too large to fit on DVD.

This change will cause problems if the collective files in a directory
are too large (that is, they are not split into subdirectories). The
script will continually try to break down the same directory even though
it cannot.