Add moduleavail macro
I really want to smack the person who chose to send the output of
'module avail' to standard error instead of standard output.

Why? WHY??????

Standard error is for reporting errors. The output of module avail is not a
f*cking error! Instead it is a standard output that is very verbose but
important.

Why do I care? Because the output of module avail almost always extends
multiple pages, so I want to redirect it to less (or more, whatever).
But the standard pipe does not redirect standard error, so I have to use
some arcane pipe redirect that I screw up all the time.

Solve the problem by making a moduleavail that modifies 'module avail' to
pipe the output correctly.