Add aliasing for github.com
It is often problematic to use the https protocol in git from behind
the Sandia firewall. This is easy to get around by using the SSH protocol
instead. For cases where loading git submodules that are specified with
the https protocol to github, automatically change them to use ssh
instead.