Add touch command to git
The touch command updates the date of the most recent commit to
the current date. Like the Unix touch command for a file, it
makes the current branch dated to the current time and it also
gives the last commit a new SHA so that it will register as
different commit but the contents will be the same.

I am putting this in so if I get an error on a GitLab dashboard
build that I think is transient, I can push a new commit that
has the same contents but will have a different SHA so I can more
easily tell which builds are new.