Put added /usr/local/bin at end of path
This path should already be in PATH. However, (I guess) sometimes
it is not. The easiest thing to do is add it (which was already
done). However, I changed it to add it to the back of the path
so the system path does not overwrite some potential custom
installed programs.